Move code around; use ParseNode::getValue
