+++ src/main/java/com/github/chrisbrenton/grappa/formal/nodes/BnfRuleDefinition.java
@@ -1,5 +1,7 @@
-package com.github.chrisbrenton.grappa.formal;
+package com.github.chrisbrenton.grappa.formal.nodes;
 
+import com.github.chrisbrenton.grappa.formal.ExpressionGenerator;
+import com.github.chrisbrenton.grappa.formal.RuleNameMangler;
 import com.github.chrisbrenton.grappa.parsetree.nodes.ParseNode;
 import com.sun.codemodel.JExpr;
 import com.sun.codemodel.JInvocation;
@@ -40,7 +42,7 @@ public final class BnfRuleDefinition
     }
 
     @Override
-    public String toString()
+    public String getValue()
     {
         return "Rule definition";
     }
