+++ src/main/java/com/github/chrisbrenton/grappa/formal/nodes/BnfRuleName.java
@@ -1,4 +1,4 @@
-package com.github.chrisbrenton.grappa.formal;
+package com.github.chrisbrenton.grappa.formal.nodes;
 
 import com.github.chrisbrenton.grappa.parsetree.nodes.ParseNode;
 
@@ -10,11 +10,5 @@ public final class BnfRuleName
     public BnfRuleName(final String value, final List<ParseNode> children)
     {
         super(value, children);
-    }
-
-    @Override
-    public String toString()
-    {
-        return getValue();
     }
 }
