+++ src/main/java/com/github/chrisbrenton/grappa/formal/nodes/BnfTerminal.java
@@ -1,5 +1,7 @@
-package com.github.chrisbrenton.grappa.formal;
+package com.github.chrisbrenton.grappa.formal.nodes;
 
+import com.github.chrisbrenton.grappa.formal.ExpressionGenerator;
+import com.github.chrisbrenton.grappa.formal.RuleNameMangler;
 import com.github.chrisbrenton.grappa.parsetree.nodes.ParseNode;
 import com.sun.codemodel.JExpr;
 import com.sun.codemodel.JExpression;
@@ -89,11 +91,5 @@ public final class BnfTerminal
         }
 
         return sb.toString();
-    }
-
-    @Override
-    public String toString()
-    {
-        return getValue();
     }
 }
