+++ src/test/java/com/github/chrisbrenton/grappa/formal/PostalAddressParsingTest.java
@@ -62,7 +62,8 @@ public final class PostalAddressParsingTest
         final ParseNode node = parseTreeListener.getRootNode();
 
         try (
-            final DotFileGenerator generator = new DotFileGenerator(outfile);
+            final DotFileGenerator generator
+                = new DotFileGenerator(outfile, ParseNode::getValue);
         ) {
             generator.render(node);
         }
